+++ CHANGELOG.md
@@ -14,6 +14,7 @@ app.
 - Make the `x` key close the popup if it is configured to _both_ close it and
   expand it.
 - Fixed recognition of words that end in half-width numerals like 小1.
+- Added parsing for ill-formed numbers like 39,800万円.
 - Made the options page show up in a new tab on Edge.
 - Added a workaround for a
   [Firefox bug](https://bugzilla.mozilla.org/show_bug.cgi?id=1860486) that would