+++ CHANGELOG.md
@@ -17,7 +17,7 @@ app.
   the popup _and_ expanding it.
 - Fixed recognition of words that end in half-width numerals like 小1.
 - Added parsing for ill-formed numbers like 39,800万円.
-- Made the options page show up in a new tab on Edge.
+- (Chrome, Edge) Made the options page show up in a new tab.
 - Added a workaround for a
   [Firefox bug](https://bugzilla.mozilla.org/show_bug.cgi?id=1860486) that would
   cause 10ten to hang when changing dictionary language.