+++ CHANGELOG.md
@@ -8,6 +8,8 @@ app.
 
 ## [Unreleased]
 
+## [1.16.0] - 2023-11-27
+
 - (Chrome, Edge) Enabled Manifest Version 3.
 - (Firefox, Thunderbird) Enabled non-persistent event pages.
 - (Chrome, Edge) Made the options page show up in a new tab.
@@ -1042,7 +1044,7 @@ app.
 
 - Initial version (yes, it took me four attempts to publish).
 
-[unreleased]: https://github.com/birchill/10ten-ja-reader/compare/v1.15.1...HEAD
+[unreleased]: https://github.com/birchill/10ten-ja-reader/compare/v1.16.0...HEAD
 [1.15.1]: https://github.com/birchill/10ten-ja-reader/compare/v1.15.0...v1.15.1
 [1.15.0]: https://github.com/birchill/10ten-ja-reader/compare/v1.14.3...v1.15.0
 [1.14.3]: https://github.com/birchill/10ten-ja-reader/compare/v1.14.1...v1.14.3
@@ -1162,3 +1164,5 @@ app.
 [0.0.6]: https://github.com/birchill/10ten-ja-reader/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/birchill/10ten-ja-reader/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/birchill/10ten-ja-reader/releases/tag/v0.0.4
+
+[1.16.0]: https://github.com/birchill/10ten-ja-reader/releases/tag/v1.16.0